#8 Refactor Create/Edit Deck Component
